Refactor some external commands in favor of built in ones
This commit changes the way killing is done.

With the line
`{ pending_git_status=$(write >&3 3>&- & printf "$!"); } 3>&1`
we save the PID of the subshell in a global variable that we can use in
`xterm_title_preexec` to kill the process so that the output is not
littered with git statuses. The git command itself decoupled from the
waiting command.