Split up preprompt
Instead of having preprompt having input arguments and looking at this
to decide what to do, we simply removed the possibility of updating only
the git status in the preprompt. This improves readability in exchange
for slight code duplication. Due to the way zsh handles scoping it is
probably not possible to refactor the code such that code duplication is
eliminated. The new function that only updates the git changes is called
`update_git_status`.