Do not unset git_fetch_pid
As unsetting the variable cannot be done due to the way zsh handles scoping
we simply set it to 0 instead. This commit also fixes a bug where the
process group was not killed. By prefixing the pid with a `-` we kill the
whole process group instead of only the process itself