Protect against prompt resize events
This commit closes a bug where the prompt would lose the git status on terminal
resize events due to zsh reprinting the prompt on every resize event. This is
because the git status was not actually part of the prompt. Now we save the
result of the query to gitstatus in a variable and make that part of the prompt.